requirements.txt
google-play-scraper
pandas
pillow
requests